Rename TextField to TextFieldAdapter
