+++ packages/react-hig/src/react-hig.js
@@ -24,4 +24,4 @@ export { default as IconButton } from './elements/basics/IconButton';
 export {
   default as RadioButton
 } from './adapters/FormElements/RadioButtonAdapter';
-export { default as TextField } from './adapters/TextField';
+export { default as TextField } from './adapters/TextFieldAdapter';
